+++ src/modules/form-builder/components/editable-div/__tests__/editable-div.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import { userEvent } from "@testing-library/user-event"
 import '@testing-library/jest-dom'
 
-import { BlockTypeEnum } from '../../..'
+import { BlockTypeEnum } from '../../../../forms/edit'
 import { EditableDiv } from '../editable-div'
 
 describe('EditableDiv', () => {
