--- src/modules/forms/edit/components/index.ts
@@ -1,1 +1,0 @@
-export * from "./edit-page-header";
