--- src/modules/forms/edit/enums/index.ts
@@ -1,1 +1,0 @@
-export * from "./edit-form.enum";
