--- src/modules/forms/edit/index.ts
@@ -1,2 +1,0 @@
-export * from "./edit";
-export * from "./enums/edit-form.enum";
