--- src/modules/forms/edit/store/index.ts
@@ -1,1 +1,0 @@
-export * from "./edit-form-store";
