+++ src/modules/homepage/homepage.tsx
@@ -7,11 +7,16 @@ const Homepage = () => {
     <div className="w-screen h-screen grid place-items-center">
       <h1 className="text-5xl font-semibold">Hello from RDS 👋</h1>
 
-      <div className="w-sm space-y-4">
+      <div className="gap-4 flex items-center">
         <Button asChild>
           <Link href="/login">
             Login
           </Link>
+
+        </Button>
+
+        <Button asChild>
+          <Link href="/create">Create</Link>
         </Button>
       </div>
     </div>
