+++ src/pages/create.tsx
@@ -1,9 +1,13 @@
-import { Edit, EditFormModeEnum } from "@/modules/forms/edit"
+// import { Edit, EditFormModeEnum } from "@/modules/forms/edit"
 
-const CreatePage = () => {
-    return (
-        <Edit mode={EditFormModeEnum.CREATE_WITHOUT_LOGIN} />
-    )
-}
+import { FormBuilder } from "@/modules/form-builder";
 
-export default CreatePage
+// const CreatePage = () => {
+//     return (
+//         <Edit mode={EditFormModeEnum.CREATE_WITHOUT_LOGIN} />
+//     )
+// }
+
+// export default CreatePage
+
+export default FormBuilder
